Copy table file and build script from HSC fork.
Table has been slightly modified to remove an optional dependency on gcc,
which LSST considers implicit.

I don't know how important the PBS path lookup code is in the build
configuration, but I'd rather err on the side of caution and leave it in
in case it's needed to build usefully on certain HSC clusters.